pyproject.toml
[project]
name = "velib-project"
version = "0.1.0"
description = "Pipeline de données Vélib avec Dagster et Spark."
readme = "README.md"
requires-python = "==3.10.*"
dependencies = [
    "dagster",
    "kafka-python",
    "dagster-webserver",
    "dagster-postgres",
    "pandas",
    "pyspark==3.5.3",
    "requests",
    "python-dotenv",
    "fastapi",
    "uvicorn",
    "duckdb",
    "boto3"
]

[project.scripts]
# Permet de lancer "velib_dev" pour démarrer dagster dev
velib_dev = "dagster dev -f src/definitions.py"

[tool.uv]
dev-dependencies = [
    "pytest",
    "ruff"     # Linter
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
fixable = ["ALL"]